// test/types/Errors.ts
export enum FA12Errors {
    FA12_NOT_ENOUGH_ALLOWANCE = "NotEnoughAllowance",
    FA12_NOT_ENOUGH_BALANCE = "NotEnoughBalance",
    FA12_NEGATIVE_TOTAL_SUPPLY = "Unexpected: Negative total supply",
    FA12_UNSAFE_ALLOWANCE_CHANGE = "UnsafeAllowanceChange",
    FA12_ALLOWANCE_MISMATCH = "AllowanceMismatch"
  }